Refactor (APIMF-3491): extracted dialect extensions parsing
